migration to new API (2/2)

+++ twittnuker/src/main/java/de/vanita5/twittnuker/loader/support/RetweetsOfMeLoader.java
@@ -41,7 +41,7 @@ import static de.vanita5.twittnuker.util.Utils.isFiltered;
 
 public class RetweetsOfMeLoader extends TwitterAPIStatusesLoader {
 
-	private int mTotalItemsCount;
+    private long mTotalItemsCount;
 
     public RetweetsOfMeLoader(final Context context, final long accountId, final long sinceId, final long maxId,
                               final List<ParcelableStatus> data, final String[] savedStatusesArgs,
@@ -49,7 +49,7 @@ public class RetweetsOfMeLoader extends TwitterAPIStatusesLoader {
         super(context, accountId, sinceId, maxId, data, savedStatusesArgs, tabPosition, fromUser);
 	}
 
-	public int getTotalItemsCount() {
+    public long getTotalItemsCount() {
 		return mTotalItemsCount;
 	}
